Refactor main.rs into logical modules
Split the monolithic 3007-line main.rs into 8 focused modules:
- models.rs: Data types, constants, and utility functions
- github.rs: GitHub CLI operations (issues, PRs, repos)
- git.rs: Git worktree operations and directory trust
- session.rs: Tmux session management and worktree+session creation
- hooks.rs: Unix socket event handling and Claude hook scripts
- config.rs: Configuration persistence
- app.rs: Application state struct and methods
- ui.rs: All TUI rendering (board, repo select, modals)

main.rs now contains only the entry point and event loop.
