+++ src/config.rs
@@ -1,0 +1,35 @@
+use std::fs;
+use std::path::PathBuf;
+
+use color_eyre::Result;
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize)]
+pub struct Config {
+    pub repo: String,
+}
+
+pub fn config_path() -> PathBuf {
+    dirs::config_dir()
+        .unwrap_or_else(|| PathBuf::from("."))
+        .join("roctopai")
+        .join("config.json")
+}
+
+pub fn load_config() -> Option<Config> {
+    let path = config_path();
+    let data = fs::read_to_string(path).ok()?;
+    serde_json::from_str(&data).ok()
+}
+
+pub fn save_config(repo: &str) -> Result<()> {
+    let path = config_path();
+    if let Some(parent) = path.parent() {
+        fs::create_dir_all(parent)?;
+    }
+    let config = Config {
+        repo: repo.to_string(),
+    };
+    fs::write(path, serde_json::to_string_pretty(&config)?)?;
+    Ok(())
+}
