+++ src/main.rs
@@ -1,13 +1,19 @@
-use std::collections::{HashMap, HashSet};
+mod app;
+mod config;
+mod git;
+mod github;
+mod hooks;
+mod models;
+mod session;
+mod ui;
+
+use std::collections::HashMap;
 use std::fs;
 use std::io;
-use std::io::Read as _;
-use std::os::unix::net::UnixListener;
-use std::path::PathBuf;
 use std::process::Command;
 use std::sync::mpsc;
 use std::sync::{Arc, Mutex};
-use std::time::{Duration, Instant};
+use std::time::Duration;
 
 use color_eyre::Result;
 use crossterm::{
@@ -15,1173 +21,18 @@ use crossterm::{
     terminal::{disable_raw_mode, enable_raw_mode, EnterAlternateScreen, LeaveAlternateScreen},
     ExecutableCommand,
 };
-use ratatui::{
-    layout::{Constraint, Direction, Layout, Rect},
-    style::{Color, Modifier, Style},
-    text::{Line, Span},
-    widgets::{Block, Borders, Clear, Padding, Paragraph, Wrap},
-    Frame, Terminal,
+
+use app::App;
+use config::{load_config, save_config};
+use git::{fetch_worktrees, remove_worktree};
+use github::{close_issue, create_issue, fetch_issues, fetch_prs, fetch_repos};
+use hooks::start_event_socket;
+use models::{
+    ConfirmAction, ConfirmModal, IssueModal, IssueSubmitResult, MergeStrategy, Mode,
+    RepoSelectPhase, Screen, SessionStates, REFRESH_INTERVAL, SOCKET_PATH,
 };
-use serde::{Deserialize, Serialize};
-
-const SOCKET_PATH: &str = "/tmp/roctopai-events.sock";
-
-type SessionStates = Arc<Mutex<HashMap<String, String>>>;
-
-fn start_event_socket(states: SessionStates) -> io::Result<()> {
-    let _ = fs::remove_file(SOCKET_PATH);
-    let listener = UnixListener::bind(SOCKET_PATH)?;
-
-    std::thread::spawn(move || {
-        for stream in listener.incoming() {
-            match stream {
-                Ok(mut stream) => {
-                    let mut buf = String::new();
-                    let _ = stream.read_to_string(&mut buf);
-                    if let Ok(event) = serde_json::from_str::<serde_json::Value>(&buf) {
-                        if let (Some(session), Some(status)) =
-                            (event["session"].as_str(), event["status"].as_str())
-                        {
-                            if let Ok(mut states) = states.lock() {
-                                states.insert(session.to_string(), status.to_string());
-                            }
-                        }
-                    }
-                }
-                Err(_) => break,
-            }
-        }
-    });
-
-    Ok(())
-}
-
-fn ensure_hook_script() -> std::result::Result<PathBuf, String> {
-    let config_dir = dirs::home_dir()
-        .ok_or("Could not find home directory")?
-        .join(".config/roctopai");
-    fs::create_dir_all(&config_dir).map_err(|e| format!("Failed to create config dir: {}", e))?;
-
-    let script_path = config_dir.join("event-hook.sh");
-    let script = format!(
-        r#"#!/bin/bash
-# Roctopai event hook - sends Claude session events to the Unix socket
-STATUS="$1"
-cat > /dev/null
-SESSION=$(basename "$PWD" | grep -o 'issue-[0-9]*')
-[ -z "$SESSION" ] && exit 0
-SOCKET="{socket}"
-[ -S "$SOCKET" ] || exit 0
-printf '{{"session":"%s","status":"%s"}}\n' "$SESSION" "$STATUS" | nc -w1 -U "$SOCKET" 2>/dev/null
-exit 0
-"#,
-        socket = SOCKET_PATH
-    );
-    fs::write(&script_path, &script).map_err(|e| format!("Failed to write hook script: {}", e))?;
-
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::PermissionsExt;
-        fs::set_permissions(&script_path, fs::Permissions::from_mode(0o755))
-            .map_err(|e| format!("Failed to chmod hook script: {}", e))?;
-    }
-
-    Ok(script_path)
-}
-
-fn write_worktree_hook_config(
-    worktree_path: &str,
-    hook_script: &str,
-) -> std::result::Result<(), String> {
-    let claude_dir = format!("{}/.claude", worktree_path);
-    fs::create_dir_all(&claude_dir).map_err(|e| format!("Failed to create .claude dir: {}", e))?;
-
-    let settings_path = format!("{}/.claude/settings.local.json", worktree_path);
-
-    // Read existing settings if present and merge
-    let mut settings: serde_json::Value = if let Ok(data) = fs::read_to_string(&settings_path) {
-        serde_json::from_str(&data).unwrap_or_else(|_| serde_json::json!({}))
-    } else {
-        serde_json::json!({})
-    };
-
-    let hook_config = serde_json::json!({
-        "PreToolUse": [{"hooks": [{"type": "command", "command": format!("'{}' working", hook_script), "async": true}]}],
-        "Stop": [{"hooks": [{"type": "command", "command": format!("'{}' waiting", hook_script)}]}],
-        "Notification": [{"matcher": "idle_prompt", "hooks": [{"type": "command", "command": format!("'{}' waiting", hook_script), "async": true}]}]
-    });
-
-    settings["hooks"] = hook_config;
-
-    fs::write(
-        &settings_path,
-        serde_json::to_string_pretty(&settings).unwrap(),
-    )
-    .map_err(|e| format!("Failed to write hook settings: {}", e))?;
-
-    Ok(())
-}
-
-#[derive(Clone, Copy, PartialEq)]
-enum StateFilter {
-    Open,
-    Closed,
-}
-
-impl StateFilter {
-    fn toggle(self) -> Self {
-        match self {
-            StateFilter::Open => StateFilter::Closed,
-            StateFilter::Closed => StateFilter::Open,
-        }
-    }
-
-    fn label(self) -> &'static str {
-        match self {
-            StateFilter::Open => "open",
-            StateFilter::Closed => "closed",
-        }
-    }
-}
-
-#[derive(Clone, Copy, PartialEq)]
-enum AssigneeFilter {
-    All,
-    Mine,
-}
-
-impl AssigneeFilter {
-    fn toggle(self) -> Self {
-        match self {
-            AssigneeFilter::All => AssigneeFilter::Mine,
-            AssigneeFilter::Mine => AssigneeFilter::All,
-        }
-    }
-
-    fn label(self) -> &'static str {
-        match self {
-            AssigneeFilter::All => "all",
-            AssigneeFilter::Mine => "mine",
-        }
-    }
-}
-
-struct Card {
-    id: String,
-    title: String,
-    description: String,
-    full_description: Option<String>,
-    tag: String,
-    tag_color: Color,
-    related: Vec<String>,
-    url: Option<String>,
-    pr_number: Option<u64>,
-    is_draft: Option<bool>,
-    is_merged: Option<bool>,
-}
-
-enum MergeStrategy {
-    Merge,
-}
-
-impl MergeStrategy {
-    fn flag(&self) -> &str {
-        match self {
-            MergeStrategy::Merge => "--merge",
-        }
-    }
-
-    fn label(&self) -> &str {
-        match self {
-            MergeStrategy::Merge => "merge",
-        }
-    }
-}
-
-enum ConfirmAction {
-    CloseIssue {
-        number: u64,
-    },
-    RemoveWorktree {
-        path: String,
-        branch: String,
-    },
-    KillSession {
-        name: String,
-    },
-    MergePr {
-        number: u64,
-        strategy: MergeStrategy,
-    },
-    RevertPr {
-        number: u64,
-    },
-}
-
-struct ConfirmModal {
-    message: String,
-    on_confirm: ConfirmAction,
-}
-
-#[derive(PartialEq)]
-enum Mode {
-    Normal,
-    Filtering { query: String },
-    CreatingIssue,
-    Confirming,
-}
-
-#[derive(PartialEq)]
-enum Screen {
-    RepoSelect,
-    Board,
-}
-
-#[derive(PartialEq)]
-enum RepoSelectPhase {
-    Typing,
-    Loading,
-    Picking,
-}
-
-struct RepoSelectState {
-    input: String,
-    repos: Vec<String>,
-    filtered_repos: Vec<String>,
-    selected: usize,
-    phase: RepoSelectPhase,
-    error: Option<String>,
-    filter_query: String,
-}
-
-impl RepoSelectState {
-    fn new() -> Self {
-        Self {
-            input: String::new(),
-            repos: Vec::new(),
-            filtered_repos: Vec::new(),
-            selected: 0,
-            phase: RepoSelectPhase::Typing,
-            error: None,
-            filter_query: String::new(),
-        }
-    }
-
-    fn update_filtered(&mut self) {
-        if self.filter_query.is_empty() {
-            self.filtered_repos = self.repos.clone();
-        } else {
-            self.filtered_repos = self
-                .repos
-                .iter()
-                .filter(|r| fuzzy_match(&self.filter_query, r))
-                .cloned()
-                .collect();
-        }
-        if self.selected >= self.filtered_repos.len() {
-            self.selected = if self.filtered_repos.is_empty() {
-                0
-            } else {
-                self.filtered_repos.len() - 1
-            };
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize)]
-struct Config {
-    repo: String,
-}
-
-fn config_path() -> PathBuf {
-    dirs::config_dir()
-        .unwrap_or_else(|| PathBuf::from("."))
-        .join("roctopai")
-        .join("config.json")
-}
-
-fn load_config() -> Option<Config> {
-    let path = config_path();
-    let data = fs::read_to_string(path).ok()?;
-    serde_json::from_str(&data).ok()
-}
-
-fn save_config(repo: &str) -> Result<()> {
-    let path = config_path();
-    if let Some(parent) = path.parent() {
-        fs::create_dir_all(parent)?;
-    }
-    let config = Config {
-        repo: repo.to_string(),
-    };
-    fs::write(path, serde_json::to_string_pretty(&config)?)?;
-    Ok(())
-}
-
-fn fetch_repos(owner: &str) -> std::result::Result<Vec<String>, String> {
-    let output = Command::new("gh")
-        .args([
-            "repo",
-            "list",
-            owner,
-            "--json",
-            "nameWithOwner",
-            "--limit",
-            "50",
-            "-q",
-            ".[].nameWithOwner",
-        ])
-        .output()
-        .map_err(|e| format!("Failed to run gh: {}", e))?;
-
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        return Err(format!("gh error: {}", stderr.trim()));
-    }
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    let repos: Vec<String> = stdout
-        .lines()
-        .filter(|l| !l.is_empty())
-        .map(|l| l.to_string())
-        .collect();
-
-    if repos.is_empty() {
-        return Err(format!("No repos found for '{}'", owner));
-    }
-
-    Ok(repos)
-}
-
-fn label_color(name: &str) -> Color {
-    match name.to_lowercase().as_str() {
-        s if s.contains("bug") => Color::Red,
-        s if s.contains("feature") || s.contains("enhancement") => Color::Green,
-        s if s.contains("documentation") || s.contains("docs") => Color::Blue,
-        s if s.contains("good first issue") || s.contains("help wanted") => Color::Cyan,
-        s if s.contains("duplicate") || s.contains("wontfix") || s.contains("invalid") => {
-            Color::Gray
-        }
-        s if s.contains("priority") || s.contains("critical") || s.contains("urgent") => {
-            Color::LightRed
-        }
-        _ => Color::Yellow,
-    }
-}
-
-fn fetch_issues(repo: &str, state: StateFilter, assignee: AssigneeFilter) -> Vec<Card> {
-    let mut args = vec![
-        "issue".to_string(),
-        "list".to_string(),
-        "--repo".to_string(),
-        repo.to_string(),
-        "--state".to_string(),
-        state.label().to_string(),
-        "--json".to_string(),
-        "number,title,body,labels".to_string(),
-        "--limit".to_string(),
-        "30".to_string(),
-    ];
-    if assignee == AssigneeFilter::Mine {
-        args.push("--assignee".to_string());
-        args.push("@me".to_string());
-    }
-    let output = Command::new("gh").args(&args).output();
-
-    let output = match output {
-        Ok(o) if o.status.success() => o,
-        _ => return Vec::new(),
-    };
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    let issues: Vec<serde_json::Value> = match serde_json::from_str(&stdout) {
-        Ok(v) => v,
-        Err(_) => return Vec::new(),
-    };
-
-    let mut cards: Vec<Card> = issues
-        .into_iter()
-        .map(|issue| {
-            let number = issue["number"].as_u64().unwrap_or(0);
-            let title = issue["title"].as_str().unwrap_or("").to_string();
-            let body = issue["body"].as_str().unwrap_or("").to_string();
-            let full_description = if body.is_empty() {
-                None
-            } else {
-                Some(body.clone())
-            };
-            let description = if body.len() > 80 {
-                format!("{}...", &body[..77])
-            } else if body.is_empty() {
-                "No description".to_string()
-            } else {
-                body
-            };
-
-            let labels = issue["labels"]
-                .as_array()
-                .map(|arr| {
-                    arr.iter()
-                        .filter_map(|l| l["name"].as_str().map(String::from))
-                        .collect::<Vec<_>>()
-                })
-                .unwrap_or_default();
-
-            let (tag, tag_color) = if let Some(first) = labels.first() {
-                (first.clone(), label_color(first))
-            } else {
-                ("open".to_string(), Color::Green)
-            };
-
-            Card {
-                id: format!("issue-{}", number),
-                title: format!("#{} {}", number, title),
-                description,
-                full_description,
-                tag,
-                tag_color,
-                related: Vec::new(),
-                url: None,
-                pr_number: None,
-                is_draft: None,
-                is_merged: None,
-            }
-        })
-        .collect();
-    // Reverse to show oldest first (gh returns newest first)
-    cards.reverse();
-    cards
-}
-
-fn fetch_prs(repo: &str, state: StateFilter, assignee: AssigneeFilter) -> Vec<Card> {
-    let mut args = vec![
-        "pr".to_string(),
-        "list".to_string(),
-        "--repo".to_string(),
-        repo.to_string(),
-        "--state".to_string(),
-        state.label().to_string(),
-        "--json".to_string(),
-        "number,title,body,isDraft,url,headRefName,state,mergedAt".to_string(),
-        "--limit".to_string(),
-        "30".to_string(),
-    ];
-    if assignee == AssigneeFilter::Mine {
-        args.push("--assignee".to_string());
-        args.push("@me".to_string());
-    }
-    let output = Command::new("gh").args(&args).output();
-
-    let output = match output {
-        Ok(o) if o.status.success() => o,
-        _ => return Vec::new(),
-    };
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    let prs: Vec<serde_json::Value> = match serde_json::from_str(&stdout) {
-        Ok(v) => v,
-        Err(_) => return Vec::new(),
-    };
-
-    let mut cards: Vec<Card> = prs
-        .into_iter()
-        .map(|pr| {
-            let number = pr["number"].as_u64().unwrap_or(0);
-            let title = pr["title"].as_str().unwrap_or("").to_string();
-            let body = pr["body"].as_str().unwrap_or("").to_string();
-            let is_draft = pr["isDraft"].as_bool().unwrap_or(false);
-            let url = pr["url"].as_str().unwrap_or("").to_string();
-            let branch = pr["headRefName"].as_str().unwrap_or("").to_string();
-            let is_merged = pr["mergedAt"].as_str().is_some();
-
-            let description = if body.len() > 80 {
-                format!("{}...", &body[..77])
-            } else if body.is_empty() {
-                branch.clone()
-            } else {
-                body
-            };
-
-            let (tag, tag_color) = if is_draft {
-                ("draft", Color::DarkGray)
-            } else {
-                ("ready", Color::Green)
-            };
-
-            // Link to related issue if branch is issue-N
-            let related = if let Some(num) = branch.strip_prefix("issue-") {
-                vec![format!("issue-{}", num)]
-            } else {
-                Vec::new()
-            };
-
-            Card {
-                id: format!("pr-{}", number),
-                title: format!("#{} {}", number, title),
-                description,
-                full_description: None,
-                tag: tag.to_string(),
-                tag_color,
-                related,
-                url: Some(url),
-                pr_number: Some(number),
-                is_draft: Some(is_draft),
-                is_merged: Some(is_merged),
-            }
-        })
-        .collect();
-    // Reverse to show oldest first (gh returns newest first)
-    cards.reverse();
-    cards
-}
-
-fn create_issue(repo: &str, title: &str, body: &str) -> std::result::Result<u64, String> {
-    let output = Command::new("gh")
-        .args([
-            "issue",
-            "create",
-            "--repo",
-            repo,
-            "--title",
-            title,
-            "--body",
-            body,
-            "--assignee",
-            "@me",
-        ])
-        .output()
-        .map_err(|e| format!("Failed to run gh: {}", e))?;
-
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        return Err(format!("gh error: {}", stderr.trim()));
-    }
-
-    // gh issue create outputs a URL like https://github.com/owner/repo/issues/10
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    let number = stdout
-        .trim()
-        .rsplit('/')
-        .next()
-        .and_then(|s| s.parse::<u64>().ok())
-        .ok_or_else(|| format!("Could not parse issue number from: {}", stdout.trim()))?;
-
-    Ok(number)
-}
-
-fn get_repo_name(repo: &str) -> &str {
-    repo.split('/').last().unwrap_or(repo)
-}
-
-fn fetch_worktrees() -> Vec<Card> {
-    let output = Command::new("git")
-        .args(["worktree", "list", "--porcelain"])
-        .output();
-
-    let output = match output {
-        Ok(o) if o.status.success() => o,
-        _ => return Vec::new(),
-    };
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    let mut cards = Vec::new();
-
-    for block in stdout.split("\n\n") {
-        let mut path = String::new();
-        let mut branch = String::new();
-        let mut is_bare = false;
-
-        for line in block.lines() {
-            if let Some(p) = line.strip_prefix("worktree ") {
-                path = p.to_string();
-            } else if let Some(b) = line.strip_prefix("branch refs/heads/") {
-                branch = b.to_string();
-            } else if line == "bare" {
-                is_bare = true;
-            }
-        }
-
-        if path.is_empty() || is_bare {
-            continue;
-        }
-
-        let display_name = if branch.is_empty() {
-            path.split('/').last().unwrap_or(&path).to_string()
-        } else {
-            branch.clone()
-        };
-
-        // Skip the main/master worktree — that's where we run from
-        let is_main = display_name == "main" || display_name == "master";
-        if is_main {
-            continue;
-        }
-        let tag = "branch";
-        let tag_color = Color::Yellow;
-
-        // Link issue-N worktrees to issue cards
-        let related = if let Some(num) = display_name.strip_prefix("issue-") {
-            vec![format!("issue-{}", num)]
-        } else {
-            Vec::new()
-        };
-
-        cards.push(Card {
-            id: format!("wt-{}", display_name),
-            title: display_name,
-            description: path,
-            full_description: None,
-            tag: tag.to_string(),
-            tag_color,
-            related,
-            url: None,
-            pr_number: None,
-            is_draft: None,
-            is_merged: None,
-        });
-    }
-
-    cards
-}
-
-fn close_issue(repo: &str, number: u64) -> std::result::Result<(), String> {
-    let output = Command::new("gh")
-        .args(["issue", "close", "--repo", repo, &number.to_string()])
-        .output()
-        .map_err(|e| format!("Failed to run gh: {}", e))?;
-
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        return Err(format!("gh error: {}", stderr.trim()));
-    }
-
-    Ok(())
-}
-
-fn remove_worktree(path: &str, branch: &str) -> std::result::Result<(), String> {
-    // Kill tmux session if it exists (named after branch)
-    let _ = Command::new("tmux")
-        .args(["kill-session", "-t", branch])
-        .output();
-
-    let output = Command::new("git")
-        .args(["worktree", "remove", "--force", path])
-        .output()
-        .map_err(|e| format!("Failed to run git: {}", e))?;
-
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        return Err(format!("git worktree remove error: {}", stderr.trim()));
-    }
-
-    // Delete the branch
-    let _ = Command::new("git").args(["branch", "-D", branch]).output();
-
-    Ok(())
-}
-
-fn fetch_merged_pr_branches(repo: &str) -> Vec<String> {
-    let output = Command::new("gh")
-        .args([
-            "pr",
-            "list",
-            "--repo",
-            repo,
-            "--state",
-            "merged",
-            "--json",
-            "headRefName",
-            "--limit",
-            "30",
-        ])
-        .output();
-
-    let output = match output {
-        Ok(o) if o.status.success() => o,
-        _ => return Vec::new(),
-    };
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    let prs: Vec<serde_json::Value> = match serde_json::from_str(&stdout) {
-        Ok(v) => v,
-        Err(_) => return Vec::new(),
-    };
-
-    prs.into_iter()
-        .filter_map(|pr| pr["headRefName"].as_str().map(|s| s.to_string()))
-        .collect()
-}
-
-fn cleanup_merged_worktrees(repo: &str, worktrees: &[Card]) -> Vec<String> {
-    let merged_branches = fetch_merged_pr_branches(repo);
-    if merged_branches.is_empty() {
-        return Vec::new();
-    }
-
-    let merged_set: HashSet<&str> = merged_branches.iter().map(|s| s.as_str()).collect();
-    let mut cleaned = Vec::new();
-
-    for wt in worktrees {
-        // worktree title is the branch name, description is the path
-        if merged_set.contains(wt.title.as_str())
-            && remove_worktree(&wt.description, &wt.title).is_ok()
-        {
-            cleaned.push(wt.title.clone());
-        }
-    }
-
-    cleaned
-}
-
-fn trust_directory(path: &str) -> std::result::Result<(), String> {
-    let claude_json = dirs::home_dir()
-        .ok_or("Could not find home directory")?
-        .join(".claude.json");
-
-    let mut config: serde_json::Value = if claude_json.exists() {
-        let data = fs::read_to_string(&claude_json)
-            .map_err(|e| format!("Failed to read .claude.json: {}", e))?;
-        serde_json::from_str(&data).map_err(|e| format!("Failed to parse .claude.json: {}", e))?
-    } else {
-        serde_json::json!({})
-    };
-
-    let abs_path = fs::canonicalize(path)
-        .map_err(|e| format!("Failed to resolve path: {}", e))?
-        .to_string_lossy()
-        .to_string();
-
-    let projects = config
-        .as_object_mut()
-        .ok_or("Invalid .claude.json format")?
-        .entry("projects")
-        .or_insert_with(|| serde_json::json!({}));
-
-    let project = projects
-        .as_object_mut()
-        .ok_or("Invalid projects format")?
-        .entry(&abs_path)
-        .or_insert_with(|| serde_json::json!({}));
-
-    project["hasTrustDialogAccepted"] = serde_json::json!(true);
-
-    fs::write(&claude_json, serde_json::to_string_pretty(&config).unwrap())
-        .map_err(|e| format!("Failed to write .claude.json: {}", e))?;
-
-    Ok(())
-}
-
-fn create_worktree_and_session(
-    repo: &str,
-    number: u64,
-    title: &str,
-    body: &str,
-    hook_script: Option<&str>,
-) -> std::result::Result<(), String> {
-    let repo_name = get_repo_name(repo);
-    let branch = format!("issue-{}", number);
-    let worktree_path = format!("../{}-issue-{}", repo_name, number);
-
-    // Create worktree with new branch
-    let output = Command::new("git")
-        .args(["worktree", "add", &worktree_path, "-b", &branch])
-        .output()
-        .map_err(|e| format!("Failed to run git: {}", e))?;
-
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        return Err(format!("git worktree add error: {}", stderr.trim()));
-    }
-
-    // Pre-trust the worktree directory for Claude
-    let _ = trust_directory(&worktree_path);
-
-    // Write Claude hook config for event socket integration
-    if let Some(script) = hook_script {
-        let _ = write_worktree_hook_config(&worktree_path, script);
-    }
-
-    // Auto-assign the issue to the current user
-    let _ = Command::new("gh")
-        .args([
-            "issue",
-            "edit",
-            "--repo",
-            repo,
-            &number.to_string(),
-            "--add-assignee",
-            "@me",
-        ])
-        .output();
-
-    // Create tmux session with neovim in the first pane
-    let output = Command::new("tmux")
-        .args([
-            "new-session",
-            "-d",
-            "-s",
-            &branch,
-            "-c",
-            &worktree_path,
-            "nvim",
-            ".",
-        ])
-        .output()
-        .map_err(|e| format!("Failed to create tmux session: {}", e))?;
-
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        return Err(format!("tmux error: {}", stderr.trim()));
-    }
-
-    // Build the Claude prompt and write to a temp file
-    let body_clean = if body.is_empty() {
-        "No description provided.".to_string()
-    } else {
-        body.lines()
-            .map(|l| l.trim())
-            .filter(|l| !l.is_empty())
-            .collect::<Vec<_>>()
-            .join(" ")
-    };
-
-    let prompt = format!(
-        "You are working on GitHub issue #{} for the repo {}. Title: {}. {} Please investigate the codebase and implement a solution for this issue. When you are confident the problem is solved, commit your changes and open a draft pull request with a clear title and description that explains what was changed and why. Reference the issue with 'Closes #{}' in the PR body. Use '--assignee @me' when creating the pull request to auto-assign it.",
-        number, repo, title, body_clean, number
-    );
-
-    // Write prompt to a temp file for safe shell expansion
-    let prompt_file = format!("/tmp/roctopai-prompt-{}.txt", number);
-    fs::write(&prompt_file, &prompt).map_err(|e| format!("Failed to write prompt file: {}", e))?;
-
-    // Split right pane running Claude with the prompt via sh -c
-    let shell_cmd = format!(
-        "claude \"$(cat '{}')\" --allowedTools Read,Edit,Bash",
-        prompt_file
-    );
-
-    let output = Command::new("tmux")
-        .args(["split-window", "-h", "-t", &branch, "-c", &worktree_path])
-        .output()
-        .map_err(|e| format!("Failed to split tmux pane: {}", e))?;
-
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        return Err(format!("tmux split error: {}", stderr.trim()));
-    }
-
-    // Wait for shell to initialize, then send the command
-    std::thread::sleep(std::time::Duration::from_millis(500));
-
-    let pane_target = format!("{}:.1", branch);
-    let _ = Command::new("tmux")
-        .args(["send-keys", "-t", &pane_target, "-l", &shell_cmd])
-        .output();
-    let _ = Command::new("tmux")
-        .args(["send-keys", "-t", &pane_target, "Enter"])
-        .output();
-
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        return Err(format!("tmux split error: {}", stderr.trim()));
-    }
-
-    Ok(())
-}
-
-fn fetch_sessions(socket_states: &SessionStates) -> Vec<Card> {
-    let output = Command::new("tmux")
-        .args(["list-sessions", "-F", "#{session_name}"])
-        .output();
-
-    let output = match output {
-        Ok(o) if o.status.success() => o,
-        _ => return Vec::new(),
-    };
-
-    let states = socket_states.lock().unwrap_or_else(|e| e.into_inner());
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    stdout
-        .lines()
-        .filter(|name| !name.is_empty())
-        .filter(|name| name.starts_with("issue-"))
-        .map(|name| {
-            // Use socket-derived state if available, otherwise fall back
-            // to pane content detection.
-            let claude_state = if let Some(status) = states.get(name) {
-                status.as_str()
-            } else {
-                let pane_target = format!("{}:.1", name);
-                let pane_content = Command::new("tmux")
-                    .args(["capture-pane", "-t", &pane_target, "-p"])
-                    .output()
-                    .ok()
-                    .and_then(|o| {
-                        if o.status.success() {
-                            Some(String::from_utf8_lossy(&o.stdout).to_string())
-                        } else {
-                            None
-                        }
-                    });
-
-                if let Some(content) = pane_content {
-                    let trimmed = content.trim_end();
-                    let last_lines: Vec<&str> = trimmed.lines().rev().take(5).collect();
-                    let waiting = last_lines.iter().any(|l| {
-                        let l = l.trim();
-                        l.starts_with('❯')
-                            || l.starts_with('>')
-                            || l.contains("What would you like")
-                    });
-                    if waiting {
-                        "idle"
-                    } else {
-                        "working"
-                    }
-                } else {
-                    "working"
-                }
-            };
-
-            let (tag, tag_color, description) = match claude_state {
-                "working" => ("working", Color::Green, "Claude is processing..."),
-                "waiting" => ("waiting", Color::Yellow, "Waiting for input"),
-                "idle" => ("idle", Color::Blue, "Ready for prompts"),
-                _ => ("working", Color::Green, "Claude is processing..."),
-            };
-
-            // Link to the related issue card
-            let related = vec![format!("{}", name)];
-
-            Card {
-                id: format!("session-{}", name),
-                title: name.to_string(),
-                description: description.to_string(),
-                full_description: None,
-                tag: tag.to_string(),
-                tag_color,
-                related,
-                url: None,
-                pr_number: None,
-                is_draft: None,
-                is_merged: None,
-            }
-        })
-        .collect()
-}
-
-fn attach_tmux_session(session: &str) -> std::result::Result<(), String> {
-    Command::new("tmux")
-        .args(["attach-session", "-t", session])
-        .stdin(std::process::Stdio::inherit())
-        .stdout(std::process::Stdio::inherit())
-        .stderr(std::process::Stdio::inherit())
-        .status()
-        .map_err(|e| format!("Failed to attach: {}", e))?;
-    Ok(())
-}
-
-struct IssueModal {
-    title: String,
-    body: String,
-    active_field: usize, // 0 = title, 1 = body
-    error: Option<String>,
-    submitting: bool,
-}
-
-impl IssueModal {
-    fn new() -> Self {
-        Self {
-            title: String::new(),
-            body: String::new(),
-            active_field: 0,
-            error: None,
-            submitting: false,
-        }
-    }
-}
-
-enum IssueSubmitResult {
-    Success {
-        number: u64,
-        worktree_result: std::result::Result<(), String>,
-    },
-    Error(String),
-}
-
-const REFRESH_INTERVAL: Duration = Duration::from_secs(30);
-
-struct App {
-    screen: Screen,
-    repo_select: RepoSelectState,
-    repo: String,
-    issues: Vec<Card>,
-    worktrees: Vec<Card>,
-    pull_requests: Vec<Card>,
-    sessions: Vec<Card>,
-    active_section: usize,
-    selected_card: [usize; 4],
-    mode: Mode,
-    issue_modal: Option<IssueModal>,
-    confirm_modal: Option<ConfirmModal>,
-    status_message: Option<String>,
-    last_refresh: Instant,
-    session_states: SessionStates,
-    hook_script_path: Option<String>,
-    issue_state_filter: StateFilter,
-    issue_assignee_filter: AssigneeFilter,
-    pr_state_filter: StateFilter,
-    pr_assignee_filter: AssigneeFilter,
-    issue_submit_rx: Option<mpsc::Receiver<IssueSubmitResult>>,
-    spinner_tick: usize,
-}
-
-impl App {
-    fn new(session_states: SessionStates) -> Self {
-        let hook_script_path = ensure_hook_script()
-            .ok()
-            .map(|p| p.to_string_lossy().to_string());
-        Self {
-            screen: Screen::RepoSelect,
-            repo_select: RepoSelectState::new(),
-            repo: String::new(),
-            issues: Vec::new(),
-            worktrees: Vec::new(),
-            pull_requests: Vec::new(),
-            active_section: 0,
-            selected_card: [0; 4],
-            mode: Mode::Normal,
-            issue_modal: None,
-            confirm_modal: None,
-            status_message: None,
-            sessions: Vec::new(),
-            last_refresh: Instant::now(),
-            session_states,
-            hook_script_path,
-            issue_state_filter: StateFilter::Open,
-            issue_assignee_filter: AssigneeFilter::All,
-            pr_state_filter: StateFilter::Open,
-            pr_assignee_filter: AssigneeFilter::All,
-            issue_submit_rx: None,
-            spinner_tick: 0,
-        }
-    }
-}
-
-impl App {
-    fn section_cards(&self, section: usize) -> &[Card] {
-        match section {
-            0 => &self.issues,
-            1 => &self.worktrees,
-            2 => &self.sessions,
-            3 => &self.pull_requests,
-            _ => &[],
-        }
-    }
-
-    fn section_card_count(&self, section: usize) -> usize {
-        self.section_cards(section).len()
-    }
-
-    fn clamp_selected(&mut self) {
-        let s = self.active_section;
-        let count = self.section_card_count(s);
-        if count == 0 {
-            self.selected_card[s] = 0;
-        } else if self.selected_card[s] >= count {
-            self.selected_card[s] = count - 1;
-        }
-    }
-
-    fn move_card_up(&mut self) {
-        let s = self.active_section;
-        if self.selected_card[s] > 0 {
-            self.selected_card[s] -= 1;
-        }
-    }
-
-    fn move_card_down(&mut self) {
-        let s = self.active_section;
-        let count = self.section_card_count(s);
-        if count > 0 && self.selected_card[s] < count - 1 {
-            self.selected_card[s] += 1;
-        }
-    }
-
-    fn selected_card_related_ids(&self) -> HashSet<String> {
-        let cards = self.section_cards(self.active_section);
-        let idx = self.selected_card[self.active_section];
-        let card = match cards.get(idx) {
-            Some(c) => c,
-            None => return HashSet::new(),
-        };
-
-        // Find the issue key that ties all related cards together.
-        // If this IS an issue card, its own id is the key.
-        // Otherwise, look for an issue-N entry in the related field.
-        let issue_key = if card.id.starts_with("issue-") {
-            Some(card.id.clone())
-        } else {
-            card.related
-                .iter()
-                .find(|r| r.starts_with("issue-"))
-                .cloned()
-        };
-
-        let Some(key) = issue_key else {
-            return card.related.iter().cloned().collect();
-        };
-
-        // Collect all cards across every column whose id matches the key
-        // or whose related list contains the key, excluding the selected card.
-        let mut ids = HashSet::new();
-        for section in 0..4 {
-            for c in self.section_cards(section) {
-                if c.id == card.id {
-                    continue;
-                }
-                if c.id == key || c.related.contains(&key) {
-                    ids.insert(c.id.clone());
-                }
-            }
-        }
-        ids
-    }
-
-    fn refresh_data(&mut self) {
-        self.issues = fetch_issues(
-            &self.repo,
-            self.issue_state_filter,
-            self.issue_assignee_filter,
-        );
-        self.pull_requests = fetch_prs(&self.repo, self.pr_state_filter, self.pr_assignee_filter);
-        self.worktrees = fetch_worktrees();
-
-        // Clean up worktrees and sessions for merged PRs
-        let cleaned = cleanup_merged_worktrees(&self.repo, &self.worktrees);
-        if !cleaned.is_empty() {
-            self.status_message = Some(format!("Cleaned up merged: {}", cleaned.join(", ")));
-            // Re-fetch worktrees after cleanup
-            self.worktrees = fetch_worktrees();
-        }
-
-        self.sessions = fetch_sessions(&self.session_states);
-        self.clamp_selected();
-        self.last_refresh = Instant::now();
-    }
-
-    fn enter_repo_select(&mut self) {
-        let owner = if self.repo.contains('/') {
-            self.repo.split('/').next().unwrap_or("").to_string()
-        } else {
-            String::new()
-        };
-        self.repo_select = RepoSelectState::new();
-        self.repo_select.input = owner;
-        self.screen = Screen::RepoSelect;
-    }
-}
+use session::{attach_tmux_session, create_worktree_and_session, fetch_sessions};
+use ui::{ui, ui_repo_select};
 
 fn main() -> Result<()> {
     color_eyre::install()?;
@@ -1193,7 +44,8 @@ fn main() -> Result<()> {
     enable_raw_mode()?;
     io::stdout().execute(EnterAlternateScreen)?;
 
-    let mut terminal = Terminal::new(ratatui::backend::CrosstermBackend::new(io::stdout()))?;
+    let mut terminal =
+        ratatui::Terminal::new(ratatui::backend::CrosstermBackend::new(io::stdout()))?;
     let mut app = App::new(session_states);
 
     // Load saved config
@@ -1236,7 +88,7 @@ fn main() -> Result<()> {
                             app.issue_assignee_filter,
                         );
                         app.clamp_selected();
-                        app.last_refresh = Instant::now();
+                        app.last_refresh = std::time::Instant::now();
                         app.issue_modal = None;
                         app.mode = Mode::Normal;
                         match worktree_result {
@@ -1466,7 +318,8 @@ fn main() -> Result<()> {
                                                         app.sessions =
                                                             fetch_sessions(&app.session_states);
                                                         app.clamp_selected();
-                                                        app.last_refresh = Instant::now();
+                                                        app.last_refresh =
+                                                            std::time::Instant::now();
                                                         app.status_message = Some(format!(
                                                             "Created worktree and session for issue #{}",
                                                             number
@@ -1582,7 +435,8 @@ fn main() -> Result<()> {
                                                             app.pr_assignee_filter,
                                                         );
                                                         app.clamp_selected();
-                                                        app.last_refresh = Instant::now();
+                                                        app.last_refresh =
+                                                            std::time::Instant::now();
                                                         app.status_message = Some(format!(
                                                             "PR #{} marked as ready",
                                                             number
@@ -1700,7 +554,7 @@ fn main() -> Result<()> {
                                         );
                                     }
                                     app.clamp_selected();
-                                    app.last_refresh = Instant::now();
+                                    app.last_refresh = std::time::Instant::now();
                                 }
                                 KeyCode::Char('m')
                                     if app.active_section == 0 || app.active_section == 3 =>
@@ -1722,7 +576,7 @@ fn main() -> Result<()> {
                                         );
                                     }
                                     app.clamp_selected();
-                                    app.last_refresh = Instant::now();
+                                    app.last_refresh = std::time::Instant::now();
                                 }
                                 KeyCode::Up | KeyCode::Char('k') => {
                                     app.move_card_up();
@@ -1747,7 +601,7 @@ fn main() -> Result<()> {
                                                         app.issue_assignee_filter,
                                                     );
                                                     app.clamp_selected();
-                                                    app.last_refresh = Instant::now();
+                                                    app.last_refresh = std::time::Instant::now();
                                                     app.status_message =
                                                         Some(format!("Closed issue #{}", number));
                                                 }
@@ -1764,7 +618,7 @@ fn main() -> Result<()> {
                                                     app.sessions =
                                                         fetch_sessions(&app.session_states);
                                                     app.clamp_selected();
-                                                    app.last_refresh = Instant::now();
+                                                    app.last_refresh = std::time::Instant::now();
                                                     app.status_message = Some(format!(
                                                         "Removed worktree '{}'",
                                                         branch
@@ -1785,7 +639,7 @@ fn main() -> Result<()> {
                                                     app.sessions =
                                                         fetch_sessions(&app.session_states);
                                                     app.clamp_selected();
-                                                    app.last_refresh = Instant::now();
+                                                    app.last_refresh = std::time::Instant::now();
                                                     app.status_message =
                                                         Some(format!("Killed session '{}'", name));
                                                 }
@@ -1842,7 +696,8 @@ fn main() -> Result<()> {
                                                                 app.pr_assignee_filter,
                                                             );
                                                             app.clamp_selected();
-                                                            app.last_refresh = Instant::now();
+                                                            app.last_refresh =
+                                                                std::time::Instant::now();
                                                             app.status_message = Some(format!(
                                                                 "Created revert PR for #{}",
                                                                 number
@@ -1897,7 +752,7 @@ fn main() -> Result<()> {
                                                     app.sessions =
                                                         fetch_sessions(&app.session_states);
                                                     app.clamp_selected();
-                                                    app.last_refresh = Instant::now();
+                                                    app.last_refresh = std::time::Instant::now();
                                                     app.status_message = Some(format!(
                                                         "Merged PR #{} ({})",
                                                         number,
@@ -2017,812 +872,9 @@ fn main() -> Result<()> {
     Ok(())
 }
 
-fn fuzzy_match(query: &str, target: &str) -> bool {
-    let target_lower = target.to_lowercase();
-    let mut target_chars = target_lower.chars();
-    for qc in query.to_lowercase().chars() {
-        loop {
-            match target_chars.next() {
-                Some(tc) if tc == qc => break,
-                Some(_) => continue,
-                None => return false,
-            }
-        }
-    }
-    true
-}
-
-fn card_matches(card: &Card, query: &str) -> bool {
-    fuzzy_match(query, &card.title) || fuzzy_match(query, &card.description)
-}
-
-fn ui_repo_select(frame: &mut Frame, state: &RepoSelectState) {
-    let area = frame.area();
-
-    // Center the content vertically
-    let vertical = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([
-            Constraint::Percentage(30),
-            Constraint::Min(0),
-            Constraint::Percentage(30),
-        ])
-        .split(area);
-
-    // Center horizontally
-    let horizontal = Layout::default()
-        .direction(Direction::Horizontal)
-        .constraints([
-            Constraint::Percentage(20),
-            Constraint::Percentage(60),
-            Constraint::Percentage(20),
-        ])
-        .split(vertical[1]);
-
-    let center = horizontal[1];
-
-    match state.phase {
-        RepoSelectPhase::Typing => {
-            let chunks = Layout::default()
-                .direction(Direction::Vertical)
-                .constraints([
-                    Constraint::Length(3),
-                    Constraint::Length(3),
-                    Constraint::Length(2),
-                    Constraint::Min(0),
-                ])
-                .split(center);
-
-            // Title
-            let title = Paragraph::new(Line::from(vec![Span::styled(
-                "Enter GitHub user or org:",
-                Style::default()
-                    .fg(Color::Cyan)
-                    .add_modifier(Modifier::BOLD),
-            )]))
-            .block(Block::default());
-            frame.render_widget(title, chunks[0]);
-
-            // Input field
-            let input_block = Block::default()
-                .borders(Borders::ALL)
-                .border_style(Style::default().fg(Color::White))
-                .title(" Owner ");
-            let input_text = Paragraph::new(Line::from(vec![
-                Span::styled(
-                    &state.input,
-                    Style::default()
-                        .fg(Color::White)
-                        .add_modifier(Modifier::BOLD),
-                ),
-                Span::styled("_", Style::default().fg(Color::Cyan)),
-            ]))
-            .block(input_block);
-            frame.render_widget(input_text, chunks[1]);
-
-            // Error message
-            if let Some(err) = &state.error {
-                let err_text = Paragraph::new(Line::from(vec![Span::styled(
-                    err.as_str(),
-                    Style::default().fg(Color::Red),
-                )]));
-                frame.render_widget(err_text, chunks[2]);
-            }
-
-            // Hint
-            let hint = Paragraph::new(Line::from(vec![Span::styled(
-                "Press Enter to fetch repos, Esc to go back",
-                Style::default().fg(Color::DarkGray),
-            )]));
-            frame.render_widget(hint, chunks[3]);
-        }
-        RepoSelectPhase::Loading => {
-            let loading = Paragraph::new(Line::from(vec![Span::styled(
-                "Fetching repositories...",
-                Style::default()
-                    .fg(Color::Yellow)
-                    .add_modifier(Modifier::BOLD),
-            )]))
-            .block(Block::default());
-            frame.render_widget(loading, center);
-        }
-        RepoSelectPhase::Picking => {
-            let max_visible = (center.height.saturating_sub(5)) as usize; // reserve space for header + filter
-            let list_height = if max_visible > 0 { max_visible } else { 1 };
-
-            let mut constraints = vec![
-                Constraint::Length(1), // title
-                Constraint::Length(1), // filter line
-                Constraint::Length(1), // separator
-            ];
-            for _ in 0..list_height.min(state.filtered_repos.len()) {
-                constraints.push(Constraint::Length(1));
-            }
-            constraints.push(Constraint::Min(0)); // hint at bottom
-
-            let chunks = Layout::default()
-                .direction(Direction::Vertical)
-                .constraints(constraints)
-                .split(center);
-
-            // Title
-            let title = Paragraph::new(Line::from(vec![
-                Span::styled(
-                    "Select a repository",
-                    Style::default()
-                        .fg(Color::Cyan)
-                        .add_modifier(Modifier::BOLD),
-                ),
-                Span::styled(
-                    format!("  ({} repos)", state.filtered_repos.len()),
-                    Style::default().fg(Color::DarkGray),
-                ),
-            ]));
-            frame.render_widget(title, chunks[0]);
-
-            // Filter line
-            let filter_line = if state.filter_query.is_empty() {
-                Paragraph::new(Line::from(vec![Span::styled(
-                    "Type to filter...",
-                    Style::default().fg(Color::DarkGray),
-                )]))
-            } else {
-                Paragraph::new(Line::from(vec![
-                    Span::styled("/ ", Style::default().fg(Color::Cyan)),
-                    Span::styled(
-                        &state.filter_query,
-                        Style::default()
-                            .fg(Color::White)
-                            .add_modifier(Modifier::BOLD),
-                    ),
-                    Span::styled("_", Style::default().fg(Color::Cyan)),
-                ]))
-            };
-            frame.render_widget(filter_line, chunks[1]);
-
-            // Separator
-            let sep = Paragraph::new(Line::from(vec![Span::styled(
-                "─".repeat(center.width as usize),
-                Style::default().fg(Color::DarkGray),
-            )]));
-            frame.render_widget(sep, chunks[2]);
-
-            // Scrolled repo list
-            let scroll_offset = if state.selected >= list_height {
-                state.selected - list_height + 1
-            } else {
-                0
-            };
-
-            let visible_count = list_height.min(state.filtered_repos.len());
-            for i in 0..visible_count {
-                let repo_idx = scroll_offset + i;
-                if repo_idx >= state.filtered_repos.len() {
-                    break;
-                }
-                let is_selected = repo_idx == state.selected;
-                let repo_name = &state.filtered_repos[repo_idx];
-                let line = if is_selected {
-                    Line::from(vec![
-                        Span::styled(" > ", Style::default().fg(Color::Cyan)),
-                        Span::styled(
-                            repo_name.as_str(),
-                            Style::default()
-                                .fg(Color::White)
-                                .add_modifier(Modifier::BOLD),
-                        ),
-                    ])
-                } else {
-                    Line::from(vec![
-                        Span::styled("   ", Style::default()),
-                        Span::styled(repo_name.as_str(), Style::default().fg(Color::Gray)),
-                    ])
-                };
-                frame.render_widget(Paragraph::new(line), chunks[3 + i]);
-            }
-
-            // Hint at bottom
-            let hint_idx = 3 + visible_count;
-            if hint_idx < chunks.len() {
-                let hint = Paragraph::new(Line::from(vec![Span::styled(
-                    "↑/↓ navigate  Enter select  Esc back",
-                    Style::default().fg(Color::DarkGray),
-                )]));
-                frame.render_widget(hint, chunks[hint_idx]);
-            }
-        }
-    }
-}
-
-fn ui(frame: &mut Frame, app: &App) {
-    let outer = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([
-            Constraint::Length(3),
-            Constraint::Min(0),
-            Constraint::Length(2),
-        ])
-        .split(frame.area());
-
-    // Top bar — selected repository
-    let repo_block = Block::default()
-        .borders(Borders::ALL)
-        .border_style(Style::default().fg(Color::Cyan))
-        .title(" Repository ");
-    let repo_text = Paragraph::new(Line::from(vec![
-        Span::styled("  ", Style::default()),
-        Span::styled(
-            &app.repo,
-            Style::default()
-                .fg(Color::White)
-                .add_modifier(Modifier::BOLD),
-        ),
-        Span::styled("  (Enter to change)", Style::default().fg(Color::DarkGray)),
-    ]))
-    .block(repo_block);
-    frame.render_widget(repo_text, outer[0]);
-
-    // Four columns
-    let columns = Layout::default()
-        .direction(Direction::Horizontal)
-        .constraints([
-            Constraint::Percentage(25),
-            Constraint::Percentage(25),
-            Constraint::Percentage(25),
-            Constraint::Percentage(25),
-        ])
-        .split(outer[1]);
-
-    let issue_title = format!(
-        " Issues [{}|{}] ",
-        app.issue_state_filter.label(),
-        app.issue_assignee_filter.label()
-    );
-    let pr_title = format!(
-        " Pull Requests [{}|{}] ",
-        app.pr_state_filter.label(),
-        app.pr_assignee_filter.label()
-    );
-    let section_data: [(&str, Color, &[Card]); 4] = [
-        (&issue_title, Color::Red, &app.issues),
-        (" Worktrees ", Color::Yellow, &app.worktrees),
-        (" Sessions ", Color::Blue, &app.sessions),
-        (&pr_title, Color::Magenta, &app.pull_requests),
-    ];
-
-    let filter_query = match &app.mode {
-        Mode::Filtering { query } => Some(query.as_str()),
-        _ => None,
-    };
-
-    let related_ids = app.selected_card_related_ids();
-
-    for (i, (title, color, cards)) in section_data.iter().enumerate() {
-        let is_active = i == app.active_section;
-        let query = if is_active { filter_query } else { None };
-        let selected = if is_active {
-            Some(app.selected_card[i])
-        } else {
-            None
-        };
-        render_column(
-            frame,
-            columns[i],
-            title,
-            *color,
-            cards,
-            is_active,
-            query,
-            selected,
-            &related_ids,
-        );
-    }
-
-    // Bottom legend bar (two lines: global on top, area-specific on bottom)
-    let key_style = Style::default()
-        .fg(Color::White)
-        .bg(Color::Rgb(60, 60, 60))
-        .add_modifier(Modifier::BOLD);
-    let desc_style = Style::default().fg(Color::Gray);
-    let key_accent = Style::default()
-        .fg(Color::Black)
-        .bg(Color::Green)
-        .add_modifier(Modifier::BOLD);
-
-    // Top line: global actions (or mode-specific actions for non-Normal modes)
-    let mut global_spans: Vec<Span> = Vec::new();
-
-    // Status message prefix
-    if let Some(msg) = &app.status_message {
-        global_spans.push(Span::styled(
-            format!(" {} ", msg),
-            Style::default()
-                .fg(Color::Yellow)
-                .add_modifier(Modifier::BOLD),
-        ));
-        global_spans.push(Span::styled(" | ", desc_style));
-    }
-
-    let global_mode_spans: Vec<Span> = match &app.mode {
-        Mode::Normal => {
-            vec![
-                Span::styled(" q/Esc ", key_style),
-                Span::styled(" Quit ", desc_style),
-                Span::styled(" Tab/S-Tab ", key_style),
-                Span::styled(" Switch column ", desc_style),
-                Span::styled(" ↑/↓ ", key_style),
-                Span::styled(" Navigate ", desc_style),
-                Span::styled(" / ", key_style),
-                Span::styled(" Filter ", desc_style),
-                Span::styled(" Enter ", key_style),
-                Span::styled(" Change repo ", desc_style),
-                Span::styled(" R ", key_style),
-                Span::styled(" Refresh ", desc_style),
-            ]
-        }
-        Mode::Filtering { .. } => vec![
-            Span::styled(" Esc ", key_style),
-            Span::styled(" Clear filter ", desc_style),
-            Span::styled(" ↑/↓ ", key_style),
-            Span::styled(" Navigate ", desc_style),
-        ],
-        Mode::CreatingIssue => vec![
-            Span::styled(" Esc ", key_style),
-            Span::styled(" Cancel ", desc_style),
-            Span::styled(" Tab ", key_style),
-            Span::styled(" Switch field ", desc_style),
-            Span::styled(" Ctrl+S ", key_accent),
-            Span::styled(" Submit ", desc_style),
-        ],
-        Mode::Confirming => vec![
-            Span::styled(" y ", key_accent),
-            Span::styled(" Confirm ", desc_style),
-            Span::styled(" n/Esc ", key_style),
-            Span::styled(" Cancel ", desc_style),
-        ],
-    };
-    global_spans.extend(global_mode_spans);
-
-    // Bottom line: area-specific actions
-    let section_names = ["Issues", "Worktrees", "Sessions", "Pull Requests"];
-    let section_colors = [Color::Red, Color::Yellow, Color::Blue, Color::Magenta];
-    let mut area_spans: Vec<Span> = Vec::new();
-
-    if matches!(app.mode, Mode::Normal) {
-        let section_label_style = Style::default()
-            .fg(section_colors[app.active_section])
-            .add_modifier(Modifier::BOLD);
-        area_spans.push(Span::styled(
-            format!(" {} ", section_names[app.active_section]),
-            section_label_style,
-        ));
-        area_spans.push(Span::styled("│ ", desc_style));
-
-        match app.active_section {
-            0 => {
-                area_spans.push(Span::styled(" n ", key_accent));
-                area_spans.push(Span::styled(" New issue ", desc_style));
-                area_spans.push(Span::styled(" w ", key_accent));
-                area_spans.push(Span::styled(" Worktree+Claude ", desc_style));
-                area_spans.push(Span::styled(" d ", key_style));
-                area_spans.push(Span::styled(" Close issue ", desc_style));
-                area_spans.push(Span::styled(" s ", key_style));
-                area_spans.push(Span::styled(" Open/Closed ", desc_style));
-                area_spans.push(Span::styled(" m ", key_style));
-                area_spans.push(Span::styled(" Assigned to me ", desc_style));
-            }
-            1 => {
-                area_spans.push(Span::styled(" v ", key_accent));
-                area_spans.push(Span::styled(" Verify (cargo run) ", desc_style));
-                area_spans.push(Span::styled(" d ", key_style));
-                area_spans.push(Span::styled(" Remove worktree ", desc_style));
-            }
-            2 => {
-                area_spans.push(Span::styled(" a ", key_accent));
-                area_spans.push(Span::styled(" Attach session ", desc_style));
-                area_spans.push(Span::styled(" d ", key_style));
-                area_spans.push(Span::styled(" Kill session ", desc_style));
-            }
-            3 => {
-                area_spans.push(Span::styled(" o ", key_accent));
-                area_spans.push(Span::styled(" Open in browser ", desc_style));
-                area_spans.push(Span::styled(" r ", key_accent));
-                area_spans.push(Span::styled(" Mark ready ", desc_style));
-                area_spans.push(Span::styled(" M ", key_accent));
-                area_spans.push(Span::styled(" Merge ", desc_style));
-                area_spans.push(Span::styled(" V ", key_accent));
-                area_spans.push(Span::styled(" Revert ", desc_style));
-                area_spans.push(Span::styled(" s ", key_style));
-                area_spans.push(Span::styled(" Open/Closed ", desc_style));
-                area_spans.push(Span::styled(" m ", key_style));
-                area_spans.push(Span::styled(" Assigned to me ", desc_style));
-            }
-            _ => {}
-        }
-    }
-
-    // Split bottom area into two rows
-    let bottom_rows = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([Constraint::Length(1), Constraint::Length(1)])
-        .split(outer[2]);
-
-    // Top row: global actions with timer on right
-    let top_bottom = Layout::default()
-        .direction(Direction::Horizontal)
-        .constraints([Constraint::Min(0), Constraint::Length(14)])
-        .split(bottom_rows[0]);
-
-    let global_legend = Paragraph::new(Line::from(global_spans));
-    frame.render_widget(global_legend, top_bottom[0]);
-
-    // Refresh countdown timer
-    let remaining = REFRESH_INTERVAL
-        .checked_sub(app.last_refresh.elapsed())
-        .unwrap_or(Duration::ZERO);
-    let secs = remaining.as_secs();
-    let timer_text = format!(" ⏱ {}s ", secs);
-    let timer_style = if secs <= 5 {
-        Style::default()
-            .fg(Color::Yellow)
-            .add_modifier(Modifier::BOLD)
-    } else {
-        Style::default().fg(Color::DarkGray)
-    };
-    let timer = Paragraph::new(Line::from(Span::styled(timer_text, timer_style)))
-        .alignment(ratatui::layout::Alignment::Right);
-    frame.render_widget(timer, top_bottom[1]);
-
-    // Bottom row: area-specific actions
-    let area_legend = Paragraph::new(Line::from(area_spans));
-    frame.render_widget(area_legend, bottom_rows[1]);
-
-    // Render issue modal overlay if open
-    if let Some(modal) = &app.issue_modal {
-        ui_issue_modal(frame, modal, app.spinner_tick);
-    }
-
-    // Render confirm modal overlay if open
-    if let Some(modal) = &app.confirm_modal {
-        ui_confirm_modal(frame, modal);
-    }
-}
-
-fn centered_rect(percent_x: u16, percent_y: u16, area: Rect) -> Rect {
-    let vertical = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([
-            Constraint::Percentage((100 - percent_y) / 2),
-            Constraint::Percentage(percent_y),
-            Constraint::Percentage((100 - percent_y) / 2),
-        ])
-        .split(area);
-    Layout::default()
-        .direction(Direction::Horizontal)
-        .constraints([
-            Constraint::Percentage((100 - percent_x) / 2),
-            Constraint::Percentage(percent_x),
-            Constraint::Percentage((100 - percent_x) / 2),
-        ])
-        .split(vertical[1])[1]
-}
-
-const SPINNER_FRAMES: &[&str] = &["⠋", "⠙", "⠹", "⠸", "⠼", "⠴", "⠦", "⠧", "⠇", "⠏"];
-
-fn ui_issue_modal(frame: &mut Frame, modal: &IssueModal, spinner_tick: usize) {
-    let area = centered_rect(50, 50, frame.area());
-
-    frame.render_widget(Clear, area);
-
-    let outer_block = Block::default()
-        .borders(Borders::ALL)
-        .border_style(Style::default().fg(Color::Cyan))
-        .title(" New Issue ")
-        .title_style(
-            Style::default()
-                .fg(Color::Black)
-                .bg(Color::Cyan)
-                .add_modifier(Modifier::BOLD),
-        )
-        .padding(Padding::new(1, 1, 1, 0));
-    let inner = outer_block.inner(area);
-    frame.render_widget(outer_block, area);
-
-    // Layout: title field (3), body field (remaining), error/spinner (1), hint (1)
-    let has_error = modal.error.is_some();
-    let has_status = has_error || modal.submitting;
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([
-            Constraint::Length(3),                              // title input
-            Constraint::Min(3),                                 // body input
-            Constraint::Length(if has_status { 1 } else { 0 }), // error or spinner
-            Constraint::Length(1),                              // hint
-        ])
-        .split(inner);
-
-    // Title field
-    let title_style = if modal.submitting {
-        Style::default().fg(Color::DarkGray)
-    } else if modal.active_field == 0 {
-        Style::default()
-            .fg(Color::White)
-            .add_modifier(Modifier::BOLD)
-    } else {
-        Style::default().fg(Color::DarkGray)
-    };
-    let title_block = Block::default()
-        .borders(Borders::ALL)
-        .border_style(title_style)
-        .title(" Title ");
-    let title_text = Paragraph::new(Line::from(vec![
-        Span::styled(
-            &modal.title,
-            Style::default().fg(if modal.submitting {
-                Color::DarkGray
-            } else {
-                Color::White
-            }),
-        ),
-        if modal.active_field == 0 && !modal.submitting {
-            Span::styled("_", Style::default().fg(Color::Cyan))
-        } else {
-            Span::raw("")
-        },
-    ]))
-    .block(title_block);
-    frame.render_widget(title_text, chunks[0]);
-
-    // Body field
-    let body_style = if modal.submitting {
-        Style::default().fg(Color::DarkGray)
-    } else if modal.active_field == 1 {
-        Style::default()
-            .fg(Color::White)
-            .add_modifier(Modifier::BOLD)
-    } else {
-        Style::default().fg(Color::DarkGray)
-    };
-    let body_block = Block::default()
-        .borders(Borders::ALL)
-        .border_style(body_style)
-        .title(" Body ");
-    let mut body_text = modal.body.clone();
-    if modal.active_field == 1 && !modal.submitting {
-        body_text.push('_');
-    }
-    let body_paragraph = Paragraph::new(body_text)
-        .style(Style::default().fg(if modal.submitting {
-            Color::DarkGray
-        } else {
-            Color::White
-        }))
-        .block(body_block)
-        .wrap(Wrap { trim: false });
-    frame.render_widget(body_paragraph, chunks[1]);
-
-    // Spinner or error
-    if modal.submitting {
-        let spinner = SPINNER_FRAMES[spinner_tick % SPINNER_FRAMES.len()];
-        let spinner_text = Paragraph::new(Line::from(vec![
-            Span::styled(
-                format!("{} ", spinner),
-                Style::default()
-                    .fg(Color::Cyan)
-                    .add_modifier(Modifier::BOLD),
-            ),
-            Span::styled(
-                "Creating issue...",
-                Style::default()
-                    .fg(Color::Yellow)
-                    .add_modifier(Modifier::BOLD),
-            ),
-        ]));
-        frame.render_widget(spinner_text, chunks[2]);
-    } else if let Some(err) = &modal.error {
-        let err_text = Paragraph::new(Line::from(vec![Span::styled(
-            err.as_str(),
-            Style::default().fg(Color::Red),
-        )]));
-        frame.render_widget(err_text, chunks[2]);
-    }
-
-    // Hint
-    let hint_text = if modal.submitting {
-        "Esc: cancel"
-    } else {
-        "Tab: switch field | Ctrl+S: submit | Esc: cancel"
-    };
-    let hint = Paragraph::new(Line::from(vec![Span::styled(
-        hint_text,
-        Style::default().fg(Color::DarkGray),
-    )]));
-    frame.render_widget(hint, chunks[3]);
-}
-
-fn ui_confirm_modal(frame: &mut Frame, modal: &ConfirmModal) {
-    let area = centered_rect(50, 20, frame.area());
-
-    frame.render_widget(Clear, area);
-
-    let outer_block = Block::default()
-        .borders(Borders::ALL)
-        .border_style(Style::default().fg(Color::Red))
-        .title(" Confirm ")
-        .title_style(
-            Style::default()
-                .fg(Color::White)
-                .bg(Color::Red)
-                .add_modifier(Modifier::BOLD),
-        )
-        .padding(Padding::new(1, 1, 1, 0));
-    let inner = outer_block.inner(area);
-    frame.render_widget(outer_block, area);
-
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([Constraint::Min(1), Constraint::Length(1)])
-        .split(inner);
-
-    let message = Paragraph::new(modal.message.as_str()).style(Style::default().fg(Color::White));
-    frame.render_widget(message, chunks[0]);
-
-    let hint = Paragraph::new(Line::from(vec![
-        Span::styled(
-            "y",
-            Style::default()
-                .fg(Color::Green)
-                .add_modifier(Modifier::BOLD),
-        ),
-        Span::styled(" confirm  ", Style::default().fg(Color::DarkGray)),
-        Span::styled(
-            "n",
-            Style::default().fg(Color::Red).add_modifier(Modifier::BOLD),
-        ),
-        Span::styled(" cancel", Style::default().fg(Color::DarkGray)),
-    ]));
-    frame.render_widget(hint, chunks[1]);
-}
-
-fn render_column(
-    frame: &mut Frame,
-    area: Rect,
-    title: &str,
-    color: Color,
-    cards: &[Card],
-    is_active: bool,
-    filter_query: Option<&str>,
-    selected: Option<usize>,
-    related_ids: &HashSet<String>,
-) {
-    let border_style = if is_active {
-        Style::default()
-            .fg(Color::White)
-            .add_modifier(Modifier::BOLD)
-    } else {
-        Style::default().fg(color)
-    };
-    let col_block = Block::default()
-        .borders(Borders::ALL)
-        .border_style(border_style)
-        .title(title)
-        .title_style(if is_active {
-            Style::default()
-                .fg(Color::Black)
-                .bg(color)
-                .add_modifier(Modifier::BOLD)
-        } else {
-            Style::default().fg(color).add_modifier(Modifier::BOLD)
-        })
-        .padding(Padding::new(1, 1, 1, 0));
-    let inner = col_block.inner(area);
-    frame.render_widget(col_block, area);
-
-    // Determine content area — if filtering, reserve a line for the search input
-    let (cards_area, filter_area) = if let Some(_) = filter_query {
-        let split = Layout::default()
-            .direction(Direction::Vertical)
-            .constraints([Constraint::Length(1), Constraint::Min(0)])
-            .split(inner);
-        (split[1], Some(split[0]))
-    } else {
-        (inner, None)
-    };
-
-    // Render filter input if active
-    if let (Some(area), Some(query)) = (filter_area, filter_query) {
-        let input = Paragraph::new(Line::from(vec![
-            Span::styled("/ ", Style::default().fg(Color::Cyan)),
-            Span::styled(
-                query,
-                Style::default()
-                    .fg(Color::White)
-                    .add_modifier(Modifier::BOLD),
-            ),
-            Span::styled("_", Style::default().fg(Color::Cyan)),
-        ]));
-        frame.render_widget(input, area);
-    }
-
-    // Filter cards
-    let visible_cards: Vec<&Card> = if let Some(query) = filter_query {
-        if query.is_empty() {
-            cards.iter().collect()
-        } else {
-            cards.iter().filter(|c| card_matches(c, query)).collect()
-        }
-    } else {
-        cards.iter().collect()
-    };
-
-    let card_height = 4u16;
-    let mut constraints: Vec<Constraint> = visible_cards
-        .iter()
-        .map(|_| Constraint::Length(card_height))
-        .collect();
-    constraints.push(Constraint::Min(0));
-
-    let slots = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints(constraints)
-        .split(cards_area);
-
-    for (i, card) in visible_cards.iter().enumerate() {
-        let is_selected = selected.is_some_and(|s| s == i);
-        let is_related = !is_selected && related_ids.contains(&card.id);
-        render_card(frame, slots[i], card, is_selected, is_related);
-    }
-}
-
-fn render_card(frame: &mut Frame, area: Rect, card: &Card, is_selected: bool, is_related: bool) {
-    let border_style = if is_selected {
-        Style::default()
-            .fg(Color::Rgb(255, 200, 50))
-            .add_modifier(Modifier::BOLD)
-    } else if is_related {
-        Style::default().fg(Color::Rgb(180, 160, 100))
-    } else {
-        Style::default().fg(Color::DarkGray)
-    };
-    let card_block = Block::default()
-        .borders(Borders::ALL)
-        .border_style(border_style);
-    let inner = card_block.inner(area);
-    frame.render_widget(card_block, area);
-
-    if inner.height == 0 || inner.width == 0 {
-        return;
-    }
-
-    let lines = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([Constraint::Length(1), Constraint::Length(1)])
-        .split(inner);
-
-    // Title line with tag
-    let tag = Span::styled(
-        format!(" {} ", card.tag),
-        Style::default().fg(Color::Black).bg(card.tag_color),
-    );
-    let title = Span::styled(
-        format!(" {}", card.title),
-        Style::default()
-            .fg(Color::White)
-            .add_modifier(Modifier::BOLD),
-    );
-    frame.render_widget(Paragraph::new(Line::from(vec![tag, title])), lines[0]);
-
-    // Description
-    let desc = Paragraph::new(Span::styled(
-        &card.description,
-        Style::default().fg(Color::Gray),
-    ));
-    frame.render_widget(desc, lines[1]);
-}
-
 #[cfg(test)]
 mod tests {
-    use super::*;
+    use std::fs;
     use std::process::Command;
 
     fn tmux_available() -> bool {
